fix: changed webconfig and appconfig to the correct package version for epplus
When the downgrade was done the web.config and app.config files was not updated with the correct version from 8.0.0 to 6.1.1. This fix is for that. When the downgrade was tested on dev and local it worked because probably we still had the 8.0.0 version installed. But when i tried to use in my new computer i had problems because i only had the 6.1.1 version. That's why this fix needed to be done. The same would occur in main site when updated.

Related work items: #4368