feat: added serial number to offline tool
- Removed CADID column because they don't use it and it was hidden
- Added serial number column
- Changed layout to include serial number

PS: Important to notice that the upload data will only work correctly after we make all the changes for rename and create serial number column on the database. This works now, but it will add and get the data from the column Model Number that is called serial_number in the database today.

Related work items: #4362, #4371